DABOM-499 refactor: StubCustomerIdResolver 테스트 유틸리티로 분리
- NotificationControllerTest, WebPushControllerTest에 중복된
  StubCustomerIdResolver 내부 클래스를 common.support 패키지로 추출
- CUSTOMER_ID 상수를 StubCustomerIdResolver.CUSTOMER_ID로 통합
- 테스트 클래스에서 불필요한 import 제거
