+++ src/test/java/com/project/domain/webpush/controller/WebPushControllerTest.java
@@ -14,23 +14,16 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.core.MethodParameter;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
 
-import com.project.common.auth.aop.CustomerId;
+import com.project.common.support.StubCustomerIdResolver;
 import com.project.domain.notification.service.NotificationService;
 import com.project.domain.webpush.service.WebPushService;
 
 @ExtendWith(MockitoExtension.class)
 class WebPushControllerTest {
-
-    private static final Long CUSTOMER_ID = 1L;
 
     @Mock private WebPushService webPushService;
 
@@ -110,25 +103,5 @@ class WebPushControllerTest {
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.success").value(true))
                 .andExpect(jsonPath("$.data").doesNotExist());
-    }
-
-    /**
-     * @CustomerId 파라미터를 고정값으로 주입하는 스텁 리졸버
-     */
-    static class StubCustomerIdResolver implements HandlerMethodArgumentResolver {
-
-        @Override
-        public boolean supportsParameter(MethodParameter parameter) {
-            return parameter.hasParameterAnnotation(CustomerId.class);
-        }
-
-        @Override
-        public Object resolveArgument(
-                MethodParameter parameter,
-                ModelAndViewContainer mavContainer,
-                NativeWebRequest webRequest,
-                WebDataBinderFactory binderFactory) {
-            return CUSTOMER_ID;
-        }
     }
 }
